Don't deserialize non successful query RPCs
Right now we deserialize and fail which dumps a big error message
with lots of text.

Instead, here we return the status code and log the text.

Ideally what I wanted to do is add the text as a field and then have
the calling function log it. But anyhow drops the fields for some reason. See comment.